Add 4-player game mode (#5)
* Add additional goals/walls

* Add horizontal paddles for players 3-4

* Add game/paddle tests

* Implement player 3-4 paddle movement keys

* Fix HorizontalPaddle collision not detecting

* Make puck initializable in all directions

* Decrease line thickness

Make it easier to see the puck if it's travelling diagonally.

* Factor out paddle angle constants

* Delete unnecessary pong.iml

* Make it easier to aim for neighbors

Widen paddle collision angle from 45 degress to 60.

* Add more tests for coverage

* Format controller/game code

* Replace score ints with integer properties

* Disable goals/paddles on player elimination

* Fix duplicate collision with corner wall + disabled goal

Replace corner wall + disabled goal with a full wall encompasing the
entire length of the player's side that was eliminated. This should
eliminate a rare corner case where a collision happens between the
disabled goal and the corner wall causing it to behave weirdly.

* Implement game-over check

* Decrease lose score 10 -> 5

* Fix test

* Keep puck angles away from exactly 45 degrees

Possibility of infinite loop as the puck bounces between corners. Ensure
that the puck angle is always initialized at least 5 degrees away from
the diagonal divider lines so that this doesn't happen.

* Add comments

* Make the stage square instead of rectangular

Fairer for top/bottom sides so that each side is the same length.

* Add more tests

* Wrap FourWayPong#getVictor() in a property obj

Less wasteful.

Co-authored-by: Paul Raupach <paul.raupach@csueastbay.edu>